fix link conversion to handle anchors if present (#11)
Markdown links can have an optional anchor at the end: a link to a specific section/heading.

If I am linking to heading `#My Heading 1` in `My Note`, then in Obsidian, the link looks like this:

```
[My Heading 1](path/.../My%20Note#My%20Heading%201)
```

However, this format is not supported in Docusaurus. Anchor links in Docusaurus cannot have spaces (`%20` is replaced with `-`) and the anchors must be lowercased.

Additionally, if there is an anchor, then the current behavior to remove the `.md` file extension will not work, since the `.md` is not the end of the url part, the anchor is, so this addresses that as well.